Bug 656197 part 3. Remove UpdateState calls in BeforeSetAttr.
The UpdateState calls in BeforeSetAttr were there in case an AttributeWillChange observer examined the element state and expected it to be updated to whatever changes BeforeSetAttr made. But at this point, AttributeWillChange runs before BeforeSetAttr, and there is no code that runs between BeforeSetAttr and the subsequent UpdateState in SetAttrAndNotify/UnsetAttr that cares or could care about the state of the element. So it's safe to do no state updates in BeforeSetAttr and just do the single UpdateState we already do.