Bug 1334247 - Remove unused nsGfxButtonControlFrame::CreateFrameFor.
 CreateFrameFor only gets called for frames that get constructed by the special nsCSSFrameConstructor::CreateAnoymousFrames path

There's nothing special about buttons, so they just take the normal AddFCItemsForAnonymousContent path. That path actually asserts against a non-trivial CreateFrameFor implementation, but the assert currently doesn't fire due to the mismatch between aPossiblyLeafFrame (which is the nsGfxButtonControlFrame used to generate the anonymous content) and aFrame (which is the inner block frame passed as the container to AddFCItemsForAnonymousContent).