Bug 1377648 - Test case for reconstruction details element when removing summary element.
Currently these two test cases don't fail either on gecko or stylo even if the patches in this patch series are not applied, but will fail once the patch for bug 1374175 (i.e. not returning reconstruction damage from compute_style_difference in some cases) landed without this patch series on stylo. That means that, on stylo, reconstruction of details element on removing child summary triggered by RecreateFramesForContent in MaybeRecreateContainerForFrameRemoval has not worked well.