Bug 1395828 (part 3) - Remove AssertParserServiceIsCorrect().
It's a bit strange for the editor to distrust the parser service in this way. The double-checking seems unnecessary, especially given that it is *buggy*: it incorrectly includes `col`, `colgroup`, and `legend` as block elements, but excludes `pre`. (It must only be called in situations where these incorrectly-classified elements are not passed in, otherwise it would have triggered by now.) So this patch removes it.

The patch also removes `li` and `pre` from the IsAnyOfHTMLElements() test in HTMLEditor::NodeIsBlockStatic. Contrary to what the comment in NodeIsBlockStatic() says, they *are* identified as block elements by the parser service.