Bug 1395828 (part 4) - Remove most uses of nsIParserService/nsParserService.
It a stateless wrapper around static methods in nsHTMLTags and nsHTMLElement, and hence an unnecessary layer of indirection that just adds complexity and slowness.

This requires making nsElementTable.h an exported header, to expose the nsHTMLElement methods.

Note: We keep the function due to use in MathML which we currently do not plan to remove in UXP.