Fix IsWebComponentsEnabled checks
This patch is based on Bug 1373155, 1373677, and 1422931.

We're keeping a IsStyledByServo check in "nsDocument::IsWebComponentsEnabled JSContext" to prevent the code from ever going down a dead code path.
StyledByServo will never work anyway since we don't have or plan to support Servo and those dead code paths will eventually be cleaned from the tree, at which point we can remove and clean the various checks also.